improvements: small code improvements
