+++ src/Hl7.Fhir.Support.Poco/Hl7.Fhir.Support.Poco.csproj
@@ -12,9 +12,7 @@
     <Title>HL7 FHIR library containing support classes for POCO support in the .NET API</Title>
     <Summary>Base classes and supporting infrastructure to work with FHIR data using POCOs.</Summary>
     <Description>This is the support library for working with the POCO representation of HL7's FHIR model data</Description>
-    <PackageTags>HL7 FHIR REST parse serialize class model</PackageTags>
-    <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
-    <PackageTags>HL7;FHIR;POCO</PackageTags>
+    <PackageTags>HL7;FHIR;POCO;REST;PARSE;SERIALIZE</PackageTags>
     <AssemblyName>Hl7.Fhir.Support.Poco</AssemblyName>
   </PropertyGroup>
