+++ src/Hl7.Fhir.Support.Poco/Rest/EntryRequest.cs
@@ -31,21 +31,7 @@ namespace Hl7.Fhir.Rest
     }
 
     //Needs to be in sync with Bundle.HTTPVerbs
-    public enum HTTPVerb
-    {
-        [EnumLiteral("GET", "http://hl7.org/fhir/http-verb"), Description("GET")]
-        GET,
-        [EnumLiteral("HEAD", "http://hl7.org/fhir/http-verb"), Description("HEAD")]
-        HEAD,
-        [EnumLiteral("POST", "http://hl7.org/fhir/http-verb"), Description("POST")]
-        POST,
-        [EnumLiteral("PUT", "http://hl7.org/fhir/http-verb"), Description("PUT")]
-        PUT,
-        [EnumLiteral("DELETE", "http://hl7.org/fhir/http-verb"), Description("DELETE")]
-        DELETE,
-        [EnumLiteral("PATH", "http://hl7.org/fhir/http-verb"), Description("PATH")]
-        PATCH,
-    }
+   
     
     public enum InteractionType
     {
