+++ src/Hl7.Fhir.Support.Poco/Rest/HttpUtil.cs
@@ -263,4 +263,20 @@ namespace Hl7.Fhir.Rest
         [EnumLiteral("respond-async")]
         RespondAsync
     }
+
+    public enum HTTPVerb
+    {
+        [EnumLiteral("GET", "http://hl7.org/fhir/http-verb"), Description("GET")]
+        GET,
+        [EnumLiteral("HEAD", "http://hl7.org/fhir/http-verb"), Description("HEAD")]
+        HEAD,
+        [EnumLiteral("POST", "http://hl7.org/fhir/http-verb"), Description("POST")]
+        POST,
+        [EnumLiteral("PUT", "http://hl7.org/fhir/http-verb"), Description("PUT")]
+        PUT,
+        [EnumLiteral("DELETE", "http://hl7.org/fhir/http-verb"), Description("DELETE")]
+        DELETE,
+        [EnumLiteral("PATCH", "http://hl7.org/fhir/http-verb"), Description("PATCH")]
+        PATCH,
+    }
 }
